move playwright to commons module
