+++ tests/common/navigator.py
